requirements.txt
PySide6>=6.6.0
apscheduler>=3.10.0
bcrypt>=4.0.0
python-osc>=1.8.3
pyserial>=3.5
wakeonlan>=3.0.0
paramiko>=3.3.0
pyartnet>=1.0.0
